pyproject.toml
[project]
name = "test-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adktools>=0.1.2",
    "google-adk>=0.1.0",
    "litellm>=1.66.1",
    "pydantic>=2.11.3",
]
